feat: add confirmation to remove game on cart

+++ react/supa-games/src/components/GameListItem.js
@@ -1,9 +1,12 @@
 import { Link } from "react-router-dom";
 
 function GameListItem({ item, type, onRemove, onAdd, priceText }) {
+
   const handleRemoveClick = (event) => {
-    event.preventDefault(); 
-    onRemove(item.fkJogo); 
+    event.preventDefault();
+    if(window.confirm(`Certeza que deseja remover ${item.title}?`) === true) {
+      onRemove(item.fkJogo); 
+    } 
   };
   
   const handleAddClick = (event) => {
